split configuration into pieces
