+++ src/ca-module.hpp
@@ -21,7 +21,7 @@
 #ifndef NDNCERT_CA_MODULE_HPP
 #define NDNCERT_CA_MODULE_HPP
 
-#include "configuration.hpp"
+#include "detail/ca-configuration.hpp"
 #include "detail/crypto-helpers.hpp"
 #include "detail/ca-storage.hpp"
